refactor: update Sidebar and LinkCard components to use dynamic styling and improve text handling
